Move for loop parser
